refactor it - common behavior regular and fast sync
